ai-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2
